添加git add, 未完善

+++ HWTest/ConvertFolder.java
@@ -1,6 +1,7 @@
 package HWTest;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.*;
 
 class ConvertFolder{
@@ -26,7 +27,7 @@ class ConvertFolder{
         }
     };
 
-    public static TreeObject dfs(String path){
+    public static TreeObject dfs(String path, Stage stage) throws IOException {
         File dir = new File(path);
         File[] fs = dir.listFiles();
         Arrays.sort(fs,comparatorFile);
@@ -35,11 +36,16 @@ class ConvertFolder{
         for(int i=0;i<fs.length;i++){
             if(fs[i].isFile()){
                 BlobObject blob = new BlobObject(fs[i]);
-                ObjectStore.add(blob);
-                blobs.add(blob);
+                if(stage.inIndex(blob.getKey())){
+                    ObjectStore.add(blob);
+                    blobs.add(blob);
+                }
             }
             if(fs[i].isDirectory()){
-                TreeObject tree = dfs(path+File.separator+fs[i].getName());
+                TreeObject tree = genTree(path+File.separator+fs[i].getName(),stage);
+                if(!stage.inIndex(tree.getKey()))
+                    return null;
+                dfs(path+File.separator+fs[i].getName(),stage);
                 trees.add(tree);
             }
         }
@@ -57,8 +63,42 @@ class ConvertFolder{
         ObjectStore.add(tree);
         return tree;
     }
+
+    public static TreeObject genTree(String path, Stage stage) throws IOException {
+        File dir = new File(path);
+        File[] fs = dir.listFiles();
+        Arrays.sort(fs,comparatorFile);
+        Vector<BlobObject> blobs = new Vector<>();
+        Vector<TreeObject> trees = new Vector<>();
+        for(int i=0;i<fs.length;i++){
+            if(fs[i].isFile()){
+                BlobObject blob = new BlobObject(fs[i]);
+                if(stage.inIndex(blob.getKey())){
+                    ObjectStore.add(blob);
+                    blobs.add(blob);
+                }
+            }
+            if(fs[i].isDirectory()){
+                TreeObject tree = genTree(path+File.separator+fs[i].getName(),stage);
+                trees.add(tree);
+            }
+        }
+        BlobObject[] blobs_array = new BlobObject[blobs.size()];
+        for(int j=0;j<blobs.size();j++){
+            blobs_array[j] = blobs.get(j);
+        }
+        TreeObject[] trees_array = new TreeObject[trees.size()];
+        for(int j=0;j<trees.size();j++){
+            trees_array[j] = trees.get(j);
+        }
+        Arrays.sort(blobs_array,comparatorObject);
+        Arrays.sort(trees_array,comparatorObject);
+        TreeObject tree = new TreeObject(dir.getName(), blobs_array, trees_array);
+        return tree;
+    }
+
     /**获取工作区目录*/
-    public static TreeObject changeFile(String path,TreeObject curTree){
+    public static TreeObject changeFile(String path,TreeObject curTree, Stage stage) throws IOException {
         //获得当前commit中树的tree,blob,存成哈希表待恢复列表
         Map<String,TreeObject> treesKey=new HashMap<String,TreeObject>();
         Map<String,BlobObject> blobsKey=new HashMap<String,BlobObject>();
@@ -78,7 +118,7 @@ class ConvertFolder{
         for(int i=0;i<fs.length;i++){
             if(fs[i].isFile()){
                 BlobObject blob = new   BlobObject(fs[i]);
-                if(!blobsKey.containsKey(blob.getKey())){
+                if(!blobsKey.containsKey(blob.getKey()) && stage.inIndex(blob.getKey())){
                     deleteFolder(fs[i].getPath());
                 }
                 else{
@@ -92,11 +132,13 @@ class ConvertFolder{
                     System.out.println(s);
                 }
                 if(!treesKey.containsKey(treeName)){
-                    deleteFolder(fs[i].getPath());
+                    TreeObject tree = genTree(path+File.separator+fs[i].getName(),stage);
+                    if(!stage.inIndex(tree.getKey()))
+                        deleteFolder(fs[i].getPath());
                 }
                 else{
                     TreeObject target=treesKey.get(treeName);
-                    TreeObject tree = changeFile(path+File.separator+fs[i].getName(),target);
+                    TreeObject tree = changeFile(path+File.separator+fs[i].getName(),target, stage);
                     treesKey.remove(tree.getKey());
                     trees.add(tree);
                 }
@@ -113,7 +155,7 @@ class ConvertFolder{
             if (!dirFile.isDirectory()){
                 dirFile.mkdir();
             }
-            TreeObject tr = changeFile(path+File.separator+tree.getDirName(),tree);
+            TreeObject tr = changeFile(path+File.separator+tree.getDirName(),tree, stage);
             trees.add(tree);
         }
         //返回递归上一层文件夹的object
@@ -190,34 +232,5 @@ class ConvertFolder{
         } else {
             return false;
         }
-    }
-    public static void main(String[] args){
-        System.out.println("请输入一个文件夹的绝对路径：");
-        Scanner sc = new Scanner(System.in);
-        String path = sc.nextLine();
-        sc.close();
-
-        VersionController vc=new VersionController(path);
-        ObjectStore a = new ObjectStore();
-        ConvertFolder convertFolder = new ConvertFolder();
-        TreeObject tree = convertFolder.dfs(path);
-
-        System.out.println(tree.getDirName());
-        System.out.println("------------");
-        System.out.println("Blobs:");
-        BlobObject[] blobs = tree.getBlobs();
-        for(int i=0; i<blobs.length; i++){
-            System.out.println(blobs[i].getFileName());
-        }
-        System.out.println("------------");
-        System.out.println("Trees:");
-        TreeObject[] trees = tree.getTrees();
-        for(int i=0; i<trees.length; i++){
-            System.out.println(trees[i].getDirName());
-        }
-        //Commit测试
-        /*Branch newHead=vc.addCommit();
-        System.out.println("------------");
-        System.out.println(newHead.getBranchName()+" "+newHead);*/
     }
 }
